Get basic birth and death working
The values probably need a lot of tweaking, but this commit gives us a
good baseline.

It is very slow once we get to 100k+ citizens. Should probably consider
removing the "days" logic, and only iterate over it once per year,
creating babies with random birthdays. It's nice to have a semblance of
time passing, though.